ローディングのスピナー表示中にスクロールをすると背景が見切れる問題に対応
CSS を使って background-color をスクロールしても適用させるのは難しかった。
そのため、component の表示非表示を切り替える方針で実装した。

それに併せて、_app.tsx にヘッダーなどを配置して、共通化も行った。
